chore(dts): split oversized declaration emitter files
